moved tagmodel to test_model
